requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
numpy
ollama
pandas
